Refactor build methods to build_utilities.py
